Created modular iOS App with Swift Packages. Created new folders corelibrary and designsystem and added new library packages to the designsystem.
