--- Teammates/CoreLibrary/CoreLibrary/CoreLibrary/CoreLibrary.h
@@ -1,18 +1,0 @@
-//
-//  CoreLibrary.h
-//  CoreLibrary
-//
-//  Created by Sachin Gurung on 8/7/24.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for CoreLibrary.
-FOUNDATION_EXPORT double CoreLibraryVersionNumber;
-
-//! Project version string for CoreLibrary.
-FOUNDATION_EXPORT const unsigned char CoreLibraryVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <CoreLibrary/PublicHeader.h>
-
-
