--- Teammates/DesignSystem/DesignSystem/DesignSystem/DesignSystem.h
@@ -1,18 +1,0 @@
-//
-//  DesignSystem.h
-//  DesignSystem
-//
-//  Created by Sachin Gurung on 8/7/24.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for DesignSystem.
-FOUNDATION_EXPORT double DesignSystemVersionNumber;
-
-//! Project version string for DesignSystem.
-FOUNDATION_EXPORT const unsigned char DesignSystemVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <DesignSystem/PublicHeader.h>
-
-
