+++ Teammates/HomeView/Sources/HomeView/HomeViewiOS.swift
@@ -1,0 +1,17 @@
+//
+//  HomeView1.swift
+//  HomeView
+//
+//  Created by Sachin Gurung on 12/26/24.
+//
+
+import SwiftUI
+
+struct HomeView: View {
+    var body: some View {
+        Text("You have no activities\nJoin some teams, and tournaments")
+                .multilineTextAlignment(.center).padding()
+    }
+}
+
+
