+++ modules/designsystem/GroupView/Sources/GroupView/GroupView.swift
@@ -1,0 +1,14 @@
+// The Swift Programming Language
+// https://docs.swift.org/swift-book
+import SwiftUI
+
+public struct GroupView: View {
+    public init () {
+        
+    }
+    
+    public var body: some View {
+        Text("You have no groups\nCreate some groups")
+            .multilineTextAlignment(.center).padding()
+    }
+}
