+++ modules/designsystem/SettingsView/Sources/SettingsView/SettingsView.swift
@@ -1,0 +1,11 @@
+// The Swift Programming Language
+// https://docs.swift.org/swift-book
+import SwiftUI
+
+public struct SettingsView: View {
+    public init(){}
+    
+    public var body: some View {
+        Text("SettingsView")
+    }
+}
